backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==4.2.13
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.94.2
h11==0.16.0
idna==3.10
msgpack==1.1.0
pillow==11.2.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.23.2
whitenoise==6.6.0
